requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
snowflake-connector-python>=3.0.0
ratelimit>=2.2.1
cryptography>=41.0.0
